# Add fine-grained locking for file audit sink.