This replaces the exclusive mutex that protected the logging of
requests and responses. This limited the overall throughput when
file auditng is enabled as only a single goroutine at a time could
be marshalling request/response data.

This creates a new mutex which only protects write access to the
the file. The lock is only taken after the data has marshalled and
is ready for writing to the file. 

Resolves #7014

Before change:
```
$ go test -bench=.
goos: darwin
goarch: amd64
pkg: github.com/hashicorp/vault/builtin/audit/file
BenchmarkAuditFile_request-8   	   30000	     54506 ns/op
PASS
ok  	github.com/hashicorp/vault/builtin/audit/file	2.288s
```

After change:
```
$ go test -bench=.
goos: darwin
goarch: amd64
pkg: github.com/hashicorp/vault/builtin/audit/file
BenchmarkAuditFile_request-8   	  100000	     15289 ns/op
PASS
ok  	github.com/hashicorp/vault/builtin/audit/file	1.767s
```